--- pyproject.toml
@@ -1,14 +1,0 @@
-[project]
-name = "movie-script-generator-agent"
-version = "0.1.0"
-description = "Add your description here"
-readme = "README.md"
-requires-python = ">=3.11"
-dependencies = [
-    "jupyter>=1.1.1",
-    "notebook>=7.4.3",
-    "pandas>=2.2.3",
-    "plotly>=6.1.2",
-    "python-dotenv>=1.1.0",
-    "requests>=2.32.3",
-]
